ediçao
potreopgk

+++ src/main/kotlin/entidades/CaixaDAgua.kt
@@ -2,18 +2,17 @@ package org.example.entidades
 
 import org.example.enumeradores.Color
 import org.example.enumeradores.Material
-import java.math.BigDecimal
 
-class CaixaDAgua (// contrustores da classe
-    val capacidade : Int,
-    val cor : Color,
-    val peso : Double,
-    val preco : BigDecimal,
-    val altura : Double,
-    val profundidade : Double,
-    val Largura : Double,
-    val tamanho : String,
-    val material : Material,
-    val conectores : Int,
+class CaixaDAgua(// contrustores da classe
+    val capacidade: Double,
+    val cor: Color,
+    val peso: Int,
+    val preco: Int,
+    val altura: Int,
+    val profundidade: Int,
+    val Largura: Double,
+    val tamanho: String,
+    val material: Material,
+    val conectores: Int,
 
     )
